pyproject.toml
[tool.poetry]
name = "equipments"
version = "0.1.0"
description = ""
authors = ["Chaus <zoneofchaos@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.6.4"
fastapi = "^0.110.0"
motor = "^3.3.2"
uvicorn = "^0.28.1"
python-dotenv = "^1.0.1"
pymongo = "^4.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
